[TASK] Improve users overview
By introducing separate user pages for admins and groupAdmins
permission handling and complexity in the pages and controllers
is reduced. Moreover, showing only users of one group at a time
helps groupAdmins with many groups and allows easier sorting and
filtering of the users.

Filtering and sorting of user tables is also improved and updates
the user tables instantaneously.

Adding the users mail, min duplicate, and average duplicate to the
group users overview helps groupAdmins during the laboratory approval
process.