[TASK] Refactor action confirmation
Providing a general confirm action component with callback events allows reuse and avoids unnecessary backend calls and page redirects.

Avoiding the backend calls in the first place also eliminates file access on the backend to gather additional information such as a human readable name for an entity by its id which is already available in the frontend.